requirements.txt
# LLM Fine-Tuning & ONNX Export Pipeline
# Optimized for NVIDIA L40S (48GB VRAM)

# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Transformers & Fine-Tuning
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
datasets>=2.15.0

# ONNX Export & Optimization
onnx>=1.15.0
onnxruntime-gpu>=1.16.0
optimum[onnxruntime-gpu]>=1.16.0

# Quantization
auto-gptq>=0.6.0
autoawq>=0.1.8

# Monitoring & Logging
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Utilities
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0

# API & Serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Development
jupyter>=1.0.0
ipywidgets>=8.1.0
black>=23.12.0
pytest>=7.4.0
